veriloans/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cleanup==7.0.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
xlwt==1.3.0
